fix: [IOBP-1021] Show retry banner when all the other cards fetch fails (#6478)
## Short description
This PR fixes the behavior of the homepage banner in the "wallet" by
ensuring the retry banner is displayed even when all API calls of the
B&P domain fail. Previously, the retry banner was not shown if the
number of onboarded cards was zero. Now, the logic has been updated so
that if the number of cards is zero and at least one API call fails, the
retry banner will be displayed.

## List of changes proposed in this pull request
- Added a logic inside the `WalletCardsContainer.tsx` to handle this
edge-case when no card is available

## How to test
- Test the app on your physical device and check that when you open the
app, before opening the "Wallet" tab, disable your internet connection
and then tap on the "Wallet tab";
- After that, check that there is visible the backoff-retry banner.

---------

Co-authored-by: Emanuele Dall'Ara <71103219+LeleDallas@users.noreply.github.com>