fix: [IOCOM-1994,IOCOM-1995] Fix for back behaviours on FIMS Auth (#6481)
## Short description
This PR fixes some anomalies on the back behaviour for the FIMS' Auth
flow

## List of changes proposed in this pull request
- When going back from an error screen, proper deallocation of FIMS
resources was not triggered (using navigation.back instead of back saga)
- On Android, physical back button event is not handled by the
assistance modal and is propagated to the underlying screen. If this
happens on the FIMS' consents screen, the flow is interrupted. Since
this anomaly affects the entire app, the back capturing has been
disabled and the user has to rely on the UI back buttons

## How to test
Using the io-dev-api-server, check that:
- the FIMS flow can be successfully retried after an error
- the Android back button does nothing on the FIMS' Auth flow

Co-authored-by: Martino Cesari Tomba <60693085+forrest57@users.noreply.github.com>