fix: [IOBP-1082] Typo in Zendesk assistance json remote path   (#6546)
> [!NOTE]
> To test this PR [check
this](https://github.com/pagopa/io-dev-api-server/pull/450)

## Short description
This pull request update the URL path and improve the handling of
Zendesk payment categories. The most important changes include modifying
the URL for the Zendesk payment configuration and refactoring the code
for checking and adding subcategories in the payment failure support
modal.

## List of changes proposed in this pull request
- Changed the URL path in the `getZendeskPaymentConfig` object from
`/assistanceTools/payment/zendeskOutcomeMapping.json` to
`/assistanceTools/payments/zendeskOutcomeMapping.json`
- Refactored the `zendeskAssistanceLogAndStart` function to improve the
handling of the `zendeskPaymentCategory` check and the addition of
subcategories to the ticket. The new code ensures that the subcategory
is only processed if `zendeskPaymentCategory` is ready and a subcategory
is found.