feat(IT Wallet): [SIW-1630] Identification through direct integration with the CieID app (#6545)
## Short description
This PR introduces direct integration with the CieID app during the
Wallet activation phase. When the CieID app is installed on a user's
device, it will be opened directly to authenticate the user. This flow
is particularly optimized for Android, thanks to
`@pagopa/io-react-native-cieid`.

When the app is not installed, identification will be carried on through
a regular web view.

## List of changes proposed in this pull request
- Created `ItwCieIdLoginScreen` to handle the CieID app:
    - Use `openCieIdApp` on Android
    - Use `Linking` on iOS to directly open CieID
    - Fallback to a regular web view when CieID is not installed 
- Removed `ITW_ISSUANCE_REDIRECT_URI_CIE` and always used the same
redirect URI
- Create a function to serialize `Error` in failure reasons for better
debugging
- Moved a few CIE related constants to `utils/cie.ts` for better reuse

## How to test
> [!NOTE]
> Needs testing on iOS.

Test all the identification methods to ensure no regressions were
introduced.

For CieID, test the following:
- Identification with CieID installed → it should directly open the app
- Identification without CieID → it should ask for CieID credentials in
a web view