feat(IT Wallet): [SIW-1824] Show alert if wallet instance is revoked (#6547)
## Short description
This PR adds functionality to display an alert in the wallet section
when the Wallet Instance has been revoked. The alerts are displayed
based on the revocation reason.

## List of changes proposed in this pull request
- add `itwUpdateWalletInstanceStatus` in
`getStatusOrResetWalletInstance` to update the wallet instance status
- update the global state of Wallet Instance including revocation
details
- add `itwWalletInstanceStatusSelector` to retrieve WI status
- add `useItwWalletInstanceRevocationAlert` to display `Alert` in based
on the revocation reason

## How to test
Simulate or revoked via Web the wallet instance revocation and navigate
to wallet section

CERTIFICATE_REVOKED_BY_ISSUER

https://github.com/user-attachments/assets/9aa867fb-f690-4da4-8e19-a1e53298e6de

REVOKED_BY_USER

https://github.com/user-attachments/assets/24899e8b-0110-4228-8edc-8f2b6da3e7e3