feat(IT Wallet): [SIW-1638] Handle unknown credential status (#6576)
## Short description
This PR handles unexpected failures of the status attestation call. When
that happens (for instance because of network problems), it is not
possibile to determine the status of a credential. This "unknown" status
is reflected in the UI and the credential detail cannot be accessed.

## List of changes proposed in this pull request
- Added new card assets for the "Not available" status
- Explicitly return `unknown` in `getCredentialStatus` (previously an
unknown status attestation would return `valid`)
- Modified `ItwCredentialCard` to display the greyed-out card
- Added the new status in the cards DS section

## How to test
The easiest way to test this PR is by mocking the result of the
`/status` endpoint so that it returns 500.

<img
src="https://github.com/user-attachments/assets/daa56de4-04f1-493a-afdb-cba22064d3b2"
width="240" />
<img
src="https://github.com/user-attachments/assets/7a4af4ff-7bbb-4c70-b07a-6767c82eba56"
width="240" />

---------

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>