feat(IT Wallet): [SIW-1944] Remove the ongoing badge when errors occur (#6592)
## Short description
This PR update the ongoing badge behavior
## List of changes proposed in this pull request
- Reduced the badge persistence limit to 24 hours
- Removed the badge if an error (Issuer Dynamic errors only) occur

## How to test
Request the MDL with the async flow.
Mock the `/status` and `/credential` endpoints with an error and codes
`404` and `403` simulating the issue. The body should be

`{ "error": "credential_voided" }` or anything else that can be mapped
from the EC.

Try to continue the flow from the message CTA and check if the error
occurs and the badge is removed.