chore(IT Wallet): [SIW-1945] Disable screenshots and recordings on ITW screens (#6595)
## Short description
This PR prevents screenshots and screen recordings in several IT Wallet
screens. In order to work on iOS, `react-native-flag-secure-android`
(currently used) has been changed with
`react-native-screenshot-prevent`.

The implementation is the following:
- On Android it leverages the `FLAG_SECURE` option;
- On iOS it creates a hidden secure text field, that renders the screen
blank.

## List of changes proposed in this pull request
- Removed `react-native-flag-secure-android`
- Installed `react-native-screenshot-prevent` 
- Added ITW routes to `screenBlackList`

## How to test
Ensure it is not possible to take screenshots or record the content of
the screen in the following:
- CIE pin screen during identification
- Trust issuer screen 
- Credential preview and detail (including trustmark)

Be sure to **disable debug mode**, because
`isAllowedSnapshotCurrentScreen` always returns true in debug.

> [!NOTE]
> iOS does not seem to offer a direct API to disable screenshots.
`react-native-screenshot-prevent` makes use of a workaround solution, so
please test it thoroughly on iOS.