fix: [IOPAE-1688] Fix infinite scroll on the services home page (#6613)
## Short description
This PR removes the control introduced in #5849 because when the user
scrolls quickly down, the _distanceFromEnd_ value returned by the
_onEndReached_ function becomes zero, preventing the next page from
loading.
Additionally, if there was an error in the call to retrieve the
institutions, we block the call and wait for a short period before
allowing the request to be sent again.

## List of changes proposed in this pull request
- updated reducer
- removed `isLoadingPaginatedInstitutionsSelector`,
`isUpdatingPaginatedInstitutionsSelector` and
`isErrorPaginatedInstitutionsSelector`
- refactored `useInstitutionsFetcher` hook
- updated tests

## How to test
Navigate to the services home screen and scroll down quickly. Check that
the next pages load correctly.