chore [IOPID-2600] Add new content on carousel (#6608)
## Short description
In this PR, changes are made to the pictograms and texts contained in
the carousel (visible when opening the app with the user not logged in).

## Demo & A11Y
| iOS | Android | 
| - | - |
| <video
src="https://github.com/user-attachments/assets/fa4a37bd-c456-496c-8946-de8e0886563f"/>
| <video
src="https://github.com/user-attachments/assets/8cdf6613-4c09-4f31-a19c-c253cab06642"/>
|

## How to test
Run the application( both on iOS and Android device) as not logged user
and test if the carousel works as the demo video.

---------

Co-authored-by: mariateresaventura <90319761+mariateresaventura@users.noreply.github.com>