chore: [IOBP-1149] Hide download receipt button if the transaction is of cart type (#6629)
## Short description
This PR hides the "Download receipt" button inside a receipt detail page
if the transaction is a cart type. By definition a `cart type`
transaction is a receipt with more than one item in the list.

## List of changes proposed in this pull request
- Hide the download button and show the unavailable receipt banner if
`isCart`;

## How to test
- Start the dev-server;
- Open the `Payments` section;
- Open a detail of transaction marked as `Pagamento multiplo` and check
that there isn't visible the download button and there is a banner
explaining it.

## Preview


https://github.com/user-attachments/assets/8c22aeac-4e2d-48e3-9b2a-1737220ca877