feat(ITWallet): [SIW-1455] Reissuing expired or expiring eID (#6580)
## Short description
This PR handle the reissuing of eID when it is expired or expiring and
manages credential requests when the eID is expired, and the Wallet
Instance is invalid

## List of changes proposed in this pull request
- Add a cta in `ItwEidLifecycleAlert`to start eID reissuing, navigating
to `ItwIdentificationModeSelectionScreen` with param `eidReissuing` set
to `true`
- Add an optional parameter `eidReissuing` to
`ItwIdentificationModeSelectionScreen`, if present send the
`start-reissuing` event to `itwEidIssuanceMachine`
- Created a new screen `ItwPresentationEidVerificationExpiredScreen`,
displayed when requesting a credential with the expired eID and a not
valid Wallet Instance
- Navigate to `ItwPresentationEidVerificationExpiredScreen` in
`WalletCardOnboardingScreen` requesting a credential but the Wallet
Instance is not valid
- In `itwEidIssuanceMachine`created a new action `setIsReissuing` to set
the `isReissuing` context state to `true`, and added a new guard
`isReissuing` to check if we are in the reissuing scenario.
Added the `start-reissuing` event to the `Idle` state, allowing a direct
transition to the `UserIdentification` state. Enhanced the
`UserIdentification` state to handle the back button differently based
on whether the flow is in the reissuing scenario or not. Updated the
`DisplayingPreview` state to handle reissuing. If `isReissuing` is
`true`, it navigates directly to the wallet; otherwise, it transitions
to the `Success` state

## How to test
To test the changes, simulate different scenarios by mocking the
`jwt.expiration` of the stored eID credential: request the eID when it
is expiring or expired, and attempt to request a credential with an
expired eID and an invalid Wallet Instance, ensuring the correct screens
and flows are triggered

Requesting credential with WI invalid and eID expired


https://github.com/user-attachments/assets/7055bd75-4cee-47ec-b7f5-362d5dcce9ce


Reissuing expiring eID (same flow for expired eID)


https://github.com/user-attachments/assets/1c64dade-6f11-4fcd-b06d-1abcff6364f9

---------

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>