refactor: [IOBP-1099] Pass language to checkout webview api (#6634)
## Short description
This pull request introduces a new utility function to derive the
language enum from the preferred locale and updates the
`handleWalletPaymentAuthorization` function to use this utility for
displaying the correct language in the webview

## List of changes proposed in this pull request
- Added `getLanguageEnumFromPreferredLocal` in `ts/utils/locale.ts` to
map the current locale to the corresponding language enum.
- Modified handleWalletPaymentAuthorization in
`ts/features/payments/checkout/saga/networking/handleWalletPaymentAuthorization.ts`
to utilize the new `getLanguageEnumFromPreferredLocale` function for
setting the language in the request body.

## How to test
- Verify that the `getLanguageEnumFromPreferredLocale` correctly maps
the current locale to the appropriate language enum
- Ensure that the `handleWalletPaymentAuthorization` function correctly
uses this utility function to set the language in the payment
authorization request
- Test the webview to confirm that the content is displayed in the
correct language based on the preferred locale settings