chore(IT Wallet): [SIW-1975] Request app review after first viewing of driving license detail screen (#6623)
## Short description
This PR adds the ability to request a review on the app store when the
driving license detail screen is viewed for the first time and the user
returns to the wallet section.

## List of changes proposed in this pull request
- Added a new action `itwReviewRequested` in the preferences actions
store of `itw`
- Added a new state `reviewRequested` to `feature.itw.preferences`
- Added a selector `itwReviewRequestedSelector` to retrieve the
`reviewRequested` state
- Created a hook `useItwReviewRequest` to monitor the `reviewRequested`
state and request an app review if needed
- Updated `ItwWalletCardsContainer` to use the new hook

## How to test
Verify that the app review request appears only when the driving license
detail is viewed for the first time and is not triggered for other
credentials.

Test Case Scenario:

1. Create ITWallet
2. Add the European Health Insurance Card credential to the wallet.
3. Add the Driving License credential to the wallet.
4. Open the detail screen of the European Health Insurance Card
credential -> **Expected Result**: No review request is displayed.
5. Open the detail screen of the Driving License credential ->
**Expected Result**: The review request is displayed.
6. Open the detail screen of the Driving License credential again ->
**Expected Result:** No review request is displayed.


https://github.com/user-attachments/assets/638552fb-99d4-4dd8-a070-5c61c03a27f3

---------

Co-authored-by: Fabio Bombardi <16268789+shadowsheep1@users.noreply.github.com>