refactor(Cross): [IOAPPX-466] replace `peter-evans` action with `gh` (#6627)
## Short description
This PR replaces community action to manipulate PR comments by
leveraging the use of `gh` in PR title check workflows.
It uses a safe way to handle the title inside a bash script, [avoiding
injections from the PR
Title](https://docs.github.com/en/actions/security-for-github-actions/security-guides/security-hardening-for-github-actions#using-an-intermediate-environment-variable).

## How to test
Try to change the PR title to see the workflows in action in different
scenarios.