feat: [IOPLT-810] Enable `allowFontScaling` by default for all citizens (#6561)
> [!important]
> This change will apply to every part of the app, except for blocks
rendered with the legacy markdown renderer. To ensure full and wide
support for dynamic font scaling in the app, we need to merge the
following PR:
> * https://github.com/pagopa/io-app/pull/6445

## Short description
This PR enables `allowFontScaling` by default for all citizens.

## List of changes proposed in this pull request
- Update `io-app-design-system` to `4.4.5`
- Minor refactoring to `FimsSuccessBody` and other screens to fix minor
TS issues
- Remove deprecated `Itw…` component to fix TS issue
- Add a new `useDetectSmallScreen` hook to handle screens where larger
text makes interaction impossible

### Preview



https://github.com/user-attachments/assets/9d09e9c3-2de3-4510-8df1-43f3db6b3f38



### Related PRs
* https://github.com/pagopa/io-app-design-system/pull/359
* https://github.com/pagopa/io-app-design-system/pull/356
* https://github.com/pagopa/io-app-design-system/pull/348

## How to test
Increase or decrease the font size in the **Accessibility** section
(Settings app) on both iOS and Android. Check that each part of the
interface responds to the change.

---------

Co-authored-by: Emanuele Dall'Ara <71103219+LeleDallas@users.noreply.github.com>