chore: [IOPLT-946] Update the main icon app for iOS + Add support for dark and tinted appearances (#6655)
> [!important]
> The new app icon is only supported in **XCode** `16.x` and above. We
need to update it first, in order for the app distribution to work. This
change is covered by the related PR.
> Depends on https://github.com/pagopa/io-app/pull/6664 

## Short description
This PR updates the main icon app for iOS and adds support for dark and
tinted appearances. [We have already tried to update
them](https://github.com/pagopa/io-app/pull/6415), but this change
blocked the app distribution at a critical time, the launch of IT
Wallet.

## List of changes proposed in this pull request
- Add three different icon assets at `1024×1024` size for default, dark
and tinted appearances

### Preview


https://github.com/user-attachments/assets/e087bb7f-a9ee-40b2-be33-ed246a5b4c9d


## How to test
1. Run the app in the local environment
2. Go to **Customize…** option through long press on the home screen
3. Change visual appearance
4. Check if the optical alignment is correct (for visual designers only
👀)

---------

Co-authored-by: Cristiano Tofani <cri.tofani@gmail.com>