refactor(IT Wallet): [SIW-2006] Replace `ItwMarkdown` with `IOMarkdown` and improve TOS WebView  (#6663)
## Short description
This PR replaces `ItwMarkdown`, now removed, with the new `IOMarkdown`
component. In this PR there is also an improvement of the IPZS TOS
WebView.

## List of changes proposed in this pull request
- Removed `ItwMarkdown` and replaced every reference with the
`IOMarkdown` component
- Added `generateLinkRuleWithCallback` utility function which returns
the rule for the `Link` component with the ability to pass a callback to
run when the user clicks on an url.
- Refactored `twIpzsPrivacyScreen` adding the `IOScrollView` component
- Refactored `ItwPrivacyWebViewComponent` to allow auto-resize on
content loaded

## How to test
Test on both iOS and Android the issuance flows and check that
everything is displayed correctly.