feat(IDPay): [IOBP-1147] Add request help from IDPay beneficiary details screen (#6622)
## ⚠️ This PR depends on #6621 ⚠️

## Short description
This PR adds the "Chiedi aiuto" call-to-action within the beneficiary
details screen. By clicking this CTA, a support ticket is automatically
created with the preselected category and subcategory, so the user is
not prompted to choose them manually.

## List of changes proposed in this pull request
- Added the `assistanceForIdPay` flag that if it's true, it doesn't
prompt the user choosing the category manually but it will be
preselected;
- Created a custom hook `useIdPaySupportModal` that shows the zendesk
support modal and pre-select the category and subcategory for that
screen;

## How to test
- Checkout the dev-server from this PR:
https://github.com/pagopa/io-dev-api-server/pull/457;
- Run `yarn sync`;
- Go inside the expense initiative detail and tap on the info icon on
the top right corner of the screen;
- Check that there is available a `Chiedi aiuto` call-to-action that
shows a modal with the FAQ;
- Check that if you tap on the button "Open a support request", it
doesn't prompt to the user to select the category of the topic;

## Preview


https://github.com/user-attachments/assets/7255110d-1f88-4420-b4af-01bab44030ec

---------

Co-authored-by: Emanuele Dall'Ara <71103219+LeleDallas@users.noreply.github.com>