feat(IT Wallet): [SIW-1960] Scan QR code for remote presentation (#6670)
## Short description
This PR introduces the initial implementation of remote presentation,
specifically handling QR code scanning.

## List of changes proposed in this pull request
- Created navigator `ItwRemoteStackNavigator`, param list
`ItwRemoteParamsList` and the necessary `routes` for the remote
presentation feature
- Introduced a new barcode type: `ITW_REMOTE`, representing the QR code
payload data
- Implemented a `decoder` to handle the new barcode type
- Introduced a new type `ItwRemoteQRCodePayload`, representing the QR
code payload data
- Added `ItwRemoteClaimsDisclosureScreen `, which is displayed after
scanning the QR code for remote presentation. This screen triggers the
`start` event in the corresponding state machine, passing the QR code
payload to initiate the remote presentation flow. (This is a basic
implementation, primarily intended to enable navigation and start the
flow).
- Introduced a new state `QRCodeValidation` in the remote machine,
representing the QR code validation step. (This is a draft
implementation and may be revised in future developments.)
- Added a `test` in the remote machine

## How to test
Scan a valid QR code for the remote presentation and check that the
loading spinner appears, indicating that the QR code has been processed
successfully



https://github.com/user-attachments/assets/7e22fbae-9c17-494a-80f0-2340c3abc49d