fix(IT Wallet): [SIW-2003] ITW auth level tracking  (#6710)
## Short description
This PR adds a new field in the `preferences` itw store that contains
the auth level used when a WI is requested.
It also handle the updates related to that field and the correct
tracking with Mixpanel

## List of changes proposed in this pull request
- Added a new field in the `preferences` itw store called authLevel that
can be `undefined`, `L2` or `L3`
- Stored the correct auth level when requesting the WI.
- Handled the store field reset when revocating the WI.
- Updated mixpanel tracking.

## How to test
Verify that the store is filled with the `L2` value when updating from a
previous version.
Try now to revoke the WI and request it again using an `L3` level
(CIE+PIN). Check redux if it is correct.
Verify that mixpanel receives the correct information.

---------

Co-authored-by: Gianluca Spada <gianluca.spada@spindox.it>