fix: [PE-937] a11y on `CgnCard` expiry date (#6726)
## Short description
This PR fixes the a11y text on the expiry date of the CgnCard component,
replacing it with more accessible text that contains the full month and
full year.

## List of changes proposed in this pull request
- Formatted the accessibilityLabel with the full month and full year in
order to be more accessible.

## How to test
Inside the wallet home screen, open the SR on your device and check that
the CGN validity is more accessible with full month and full year as
values.