feat(IT Wallet): [SIW-1972] Remote presentation same device (#6689)
## Short description
This PR introduces the ability to start the remote presentation flow
through a deep link in a same-device scenario.

## List of changes proposed in this pull request
- Updated `useItwLinkingOptions` to include linking options for remote
deep link
- Added `ItwRemoteRequestValidationScreen`, which receives the `payload`
and validates it. If the `payload` is invalid, an error screen, for
deeplink, will be displayed (TODO). If valid, it sends the `start` event
to the remote presentation machine
- Modified `decodeItwRemoteBarcode` to validate that the URL starts with
https://continua.io.pagopa.it/itw/auth, checks the content of the QR
code against the `ItwRemoteRequestPayload` type, and finally sends the
URL


## How to test
Use the following link as a deep link and to generate a QR code:

https://continua.io.pagopa.it/itw/auth?client_id=abc123xy&request_uri=https%3A%2F%2Fexample.com%2Fcallback&state=hyqizm592

---------

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>