feat(IT Wallet): [SIW-1961]  Handle invalid payload in deeplink for Remote presentation (#6724)
## Short description
This PR handles the view for an invalid payload in the deeplink during
the remote presentation flow

> [!WARNING]  
> This PR depends on https://github.com/pagopa/io-app/pull/6689

## List of changes proposed in this pull request
- Added a new Screen `ItwRemoteDeepLinkFailureScreen`
- Added a new file, `ItwFailureTypes.ts`, specifically to manage errors
for the `support bottom sheet`, outside the state machine flows. This
type of error is intended to be global for ITWallet, and the flow to
which it belongs must be specified in the code
- Created the `ZendeskSubcategoryValue` enum in
`useItwFailureSupportModal`, representing possible Zendesk subcategory
values for opening a support ticket. This is passed as a prop when using
`useItwFailureSupportModal`

## How to test
To test the changes, use an invalid deeplink (e.g.,
https://continua.io.pagopa.it/itw/auth?client_id=abc123x) and verify
that the support bottom sheet appears.

> [!NOTE]  
> I don't have the possibility to use deeplinks during development. I
use the same deeplink flow but by scanning the QR code instead



https://github.com/user-attachments/assets/09c50fac-6471-4003-ab2b-539d8e7e05f4

---------

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>