chore(IT Wallet): [SIW-2037] Add alert with modal in offline wallet screen (#6739)
## Short description
This PR adds an alert which informs the user that the app is currently
working in offline mode. The alert can be tapped to show a bottom sheet
with more informations.

### Preview


https://github.com/user-attachments/assets/7cc78622-2ff5-4f53-ace5-56d6b4c27a03

## List of changes proposed in this pull request
- added `withOfflineAlert` HOC which allows to wrap a screen with an
`AlertEdgeToEdgeWrapper` component displaying the offline alert
- wrapped `ItwOfflineWalletScreen` with `withOfflineAlert` HOC

## How to test
Go to **Settings → Playgrounds → Documenti su IO → Offline section**,
check that the Alert is displayed correctly.

---------

Co-authored-by: Jacopo Pompilii <Jacopo.pompilii@pagopa.it>